feat: token

+++ docker_entrypoint.py
@@ -20,27 +20,37 @@ def main():
     counter = 0
 
     while True:
-        response = get(
-            "http://169.254.169.254/latest/meta-data/spot/termination-time"
-        )
-        if response.status_code == 200:
-            kube_command = ['kubectl', 'drain', node_name]
-            kube_command += drain_parameters.split()
+        sleep(15)
+        url = "http://169.254.169.254/latest/api/token"
+        headers = {"X-aws-ec2-metadata-token-ttl-seconds": "180"}
+        token_response = requests.put(url, headers=headers)
 
-            print("Draining node: %s" % node_name)
-            result = call(kube_command)
-            if result == 0:
-                print('Node Drain successful')
-                break
+        if token_response.status_code == 200:
+            token = token_response.text
+            headers = {"X-aws-ec2-metadata-token": token}
+            response = requests.get("http://169.254.169.254/latest/meta-data/spot/termination-time", headers=headers)
 
+            if response.status_code == 200:
+                kube_command = ['kubectl', 'drain', node_name]
+                kube_command += drain_parameters.split()
+
+                print("Draining node: %s" % node_name)
+                result = call(kube_command)
+                if result == 0:
+                    print('Node Drain successful')
+                    break
+            else:
+                if counter == 60:
+                    counter = 0
+                    print("Termination notice status: %s, on Node: %s" %
+                        (response.status_code, node_name)
+                        )
+                counter += 5
+                sleep(5)
         else:
-            if counter == 60:
-                counter = 0
-                print("Termination notice status: %s, on Node: %s" %
-                      (response.status_code, node_name)
-                      )
-            counter += 5
-            sleep(5)
+            print("No Token: %s" %
+                (token_response.status_code)
+                )
 
 
 if __name__ == '__main__':
